Updated delta time to track as double

+++ Hivemind/Source/Hivemind.Library.Test/FooBee.cpp
@@ -1,6 +1,10 @@
 #include "pch.h"
 #include "FooBee.h"
 
+/**
+*	@Author: Dale Diaz
+*	@Date: 7/25/2017
+*/
 
 FooBee::FooBee(const sf::Vector2f& position, Hive& hive) :
 	Bee(position, hive)
@@ -11,7 +15,7 @@ FooBee::~FooBee()
 {
 }
 
-void FooBee::Update(sf::RenderWindow& window, const float& deltaTime)
+void FooBee::Update(sf::RenderWindow& window, const double& deltaTime)
 {
 	UNREFERENCED_PARAMETER(window);
 	UNREFERENCED_PARAMETER(deltaTime);
