+++ Hivemind/Source/Hivemind.Library/CollisionNode.h
@@ -3,7 +3,7 @@
 
 /**
 *	@Author: Dale Diaz
-*	@Date: 7/21/2017
+*	@Date: 7/25/2017
 */
 
 class CollisionNode :public Entity
@@ -30,7 +30,7 @@ public:
 	 * @Param window: The window that the collision node is being rendered to
 	 * @Param deltaTime: The time since last update call
 	 */
-	void Update(sf::RenderWindow& window, const float& deltaTime) override;
+	void Update(sf::RenderWindow& window, const double& deltaTime) override;
 
 	/**
 	 * Renders the collision node to the window
