+++ Hivemind/Source/Hivemind.Library/Drone.cpp
@@ -3,7 +3,7 @@
 
 /**
 *	@Author: Dale Diaz
-*	@Date: 7/21/2017
+*	@Date: 7/25/2017
 */
 
 using namespace std;
@@ -19,7 +19,7 @@ Drone::~Drone()
 {
 }
 
-void Drone::Update(sf::RenderWindow& window, const float& deltaTime)
+void Drone::Update(sf::RenderWindow& window, const double& deltaTime)
 {
 	Bee::Update(window, deltaTime);
