+++ Hivemind/Source/Hivemind.Library/FoodSource.h
@@ -77,7 +77,7 @@ public:
 	 * @Param window: The window that the simulation is being rendered to
 	 * @Param deltaTime: The time since the last Update call
 	 */
-	void Update(sf::RenderWindow& window, const float& deltaTime) override;
+	void Update(sf::RenderWindow& window, const double& deltaTime) override;
 
 	/**
 	 * Render method called by the main game loop
